Moved metrics up to networking component
